requirements.txt
streamlit==1.45.1
torch==2.7.1
torchvision==0.22.1
pandas==2.3.0
pillow==11.2.1
numpy==2.3.1
huggingface_hub==0.22.2
evaluate==0.4.0
